misc: file arrangements
